Rename Context to NetworkContext
